weights/requirements.txt
flask
flask-cors
ultralytics
torch
torchvision
numpy
opencv-python-headless
pillow
requests
python-dotenv
groq
